Relax the loc_in_brackets switch to be a bit more fuzzy.  If "true" behave like set to "false" anyway in case the local name matches the first word in the generic name tag.
This honors the convention, that the name in the more "important" language
on the ground is also the first name in the generic name tag.

This will make sense especially in bilingual areas where our target language
should be used first (if one of them) rather than some less important
regional one.